+++ memorize.ai/View Models/ReviewViewModel.swift
@@ -186,7 +186,7 @@ final class ReviewViewModel: ViewModel {
 			completion: {
 				if gainXP {
 					self.user.documentReference.updateData([
-						"xp": FieldValue.increment(1.0)
+						"xp": FieldValue.increment(1 as Int64)
 					]) as Void
 				}
